requirements.txt
python-dotenv==1.0.0
kafka-python==2.0.2
elasticsearch==8.11.0
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
